Create new record folder and seperate out filtering tests
